+++ CHANGELOG.md
@@ -174,6 +174,8 @@ Interaction and emacs mode
 Backends
 --------
 
+* New `backendInteractTop/backendInteractHole` fields for providing backend-specific interaction commands (run with keyboard shortcut `C-c C-i`).
+
 Other issues closed
 -------------------